@@ -12,12 +12,12 @@
   },
   "dependencies": {
     "axios": "0.16.2",
-    "element-ui": "1.3.7",
-    "js-cookie": "^2.1.4",
-    "normalize.css": "3.0.2",
-    "nprogress": "^0.2.0",
-    "vue": "2.3.3",
-    "vue-router": "2.5.3",
+    "element-ui": "1.4.3",
+    "js-cookie": "2.1.4",
+    "normalize.css": "7.0.0",
+    "nprogress": "0.2.0",
+    "vue": "2.4.2",
+    "vue-router": "2.7.0",
     "vuex": "2.3.1"
   },
   "devDependencies": {
@@ -53,10 +53,11 @@
     "rimraf": "2.6.0",
     "node-sass": "^4.5.0",
     "sass-loader": "6.0.5",
+    "svg-sprite-loader": "3.2.4",
     "url-loader": "0.5.8",
-    "vue-loader": "12.1.0",
+    "vue-loader": "13.0.4",
     "vue-style-loader": "3.0.1",
-    "vue-template-compiler": "2.3.3",
+    "vue-template-compiler": "2.4.2",
     "webpack": "2.6.1",
     "webpack-dev-middleware": "1.10.0",
     "webpack-hot-middleware": "2.18.0",
