@@ -9,8 +9,8 @@ export default {
   name: 'app'
 }
 </script>
+
 <style lang="scss">
+  @import '~normalize.css/normalize.css';// normalize.css 样式格式化
   @import './styles/index.scss'; // 全局自定义的css样式
 </style>
-
-
