@@ -1,1 +1,1 @@
-module.exports = file => require('@/views/' + file + '.vue')
+module.exports = file => require('@/views/' + file + '.vue').default
