@@ -1,25 +1,24 @@
-import axios from 'axios';
-import { Message } from 'element-ui';
-import store from '../store';
-import { getToken } from '@/utils/auth';
-
+import axios from 'axios'
+import { Message } from 'element-ui'
+import store from '../store'
+import { getToken } from '@/utils/auth'
 
 // 创建axios实例
 const service = axios.create({
   baseURL: process.env.BASE_API, // api的base_url
   timeout: 15000                  // 请求超时时间
-});
+})
 
 // request拦截器
 service.interceptors.request.use(config => {
   if (store.getters.token) {
-    config.headers['X-Token'] = getToken(); // 让每个请求携带自定义token 请根据实际情况自行修改
+    config.headers['X-Token'] = getToken() // 让每个请求携带自定义token 请根据实际情况自行修改
   }
-  return config;
+  return config
 }, error => {
   // Do something with request error
-  console.log(error); // for debug
-  Promise.reject(error);
+  console.log(error) // for debug
+  Promise.reject(error)
 })
 
 // respone拦截器
@@ -28,13 +27,13 @@ service.interceptors.response.use(
   /**
   * code为非20000是抛错 可结合自己业务进行修改
   */
-    const res = response.data;
+    const res = response.data
     if (res.code !== 20000) {
       Message({
         message: res.data,
         type: 'error',
         duration: 5 * 1000
-      });
+      })
 
       // 50008:非法的token; 50012:其他客户端登录了;  50014:Token 过期了;
       if (res.code === 50008 || res.code === 50012 || res.code === 50014) {
@@ -44,24 +43,24 @@ service.interceptors.response.use(
           type: 'warning'
         }).then(() => {
           store.dispatch('FedLogOut').then(() => {
-            location.reload();// 为了重新实例化vue-router对象 避免bug
-          });
+            location.reload()// 为了重新实例化vue-router对象 避免bug
+          })
         })
       }
-      return Promise.reject('error');
+      return Promise.reject('error')
     } else {
-      return response.data;
+      return response.data
     }
   },
   error => {
-    console.log('err' + error);// for debug
+    console.log('err' + error)// for debug
     Message({
       message: error.message,
       type: 'error',
       duration: 5 * 1000
-    });
-    return Promise.reject(error);
+    })
+    return Promise.reject(error)
   }
 )
 
-export default service;
+export default service
