@@ -4,15 +4,15 @@
 
  export function parseTime(time, cFormat) {
    if (arguments.length === 0) {
-     return null;
+     return null
    }
-   const format = cFormat || '{y}-{m}-{d} {h}:{i}:{s}';
-   let date;
-   if (typeof time == 'object') {
-     date = time;
+   const format = cFormat || '{y}-{m}-{d} {h}:{i}:{s}'
+   let date
+   if (typeof time === 'object') {
+     date = time
    } else {
-     if (('' + time).length === 10) time = parseInt(time) * 1000;
-     date = new Date(time);
+     if (('' + time).length === 10) time = parseInt(time) * 1000
+     date = new Date(time)
    }
    const formatObj = {
      y: date.getFullYear(),
@@ -22,24 +22,24 @@
      i: date.getMinutes(),
      s: date.getSeconds(),
      a: date.getDay()
-   };
+   }
    const time_str = format.replace(/{(y|m|d|h|i|s|a)+}/g, (result, key) => {
-     let value = formatObj[key];
-     if (key === 'a') return ['一', '二', '三', '四', '五', '六', '日'][value - 1];
+     let value = formatObj[key]
+     if (key === 'a') return ['一', '二', '三', '四', '五', '六', '日'][value - 1]
      if (result.length > 0 && value < 10) {
-       value = '0' + value;
+       value = '0' + value
      }
-     return value || 0;
-   });
-   return time_str;
+     return value || 0
+   })
+   return time_str
  }
 
  export function formatTime(time, option) {
-   time = +time * 1000;
-   const d = new Date(time);
-   const now = Date.now();
+   time = +time * 1000
+   const d = new Date(time)
+   const now = Date.now()
 
-   const diff = (now - d) / 1000;
+   const diff = (now - d) / 1000
 
    if (diff < 30) {
      return '刚刚'
