@@ -1,18 +1,18 @@
 <template>
-    <section class="app-main">
-        <transition name="fade" mode="out-in">
-            <router-view :key="key"></router-view>
-        </transition>
-    </section>
+  <section class="app-main">
+    <transition name="fade" mode="out-in">
+      <router-view :key="key"></router-view>
+    </transition>
+  </section>
 </template>
 
 <script>
-    export default {
-      name: 'AppMain',
-      computed: {
-        key() {
-          return this.$route.name !== undefined ? this.$route.name + +new Date() : this.$route + +new Date()
-        }
-      }
+export default {
+  name: 'AppMain',
+  computed: {
+    key() {
+      return this.$route.name !== undefined ? this.$route.name + +new Date() : this.$route + +new Date()
     }
+  }
+}
 </script>
