@@ -1,31 +1,32 @@
 <template>
-    <div class="app-wrapper" :class="{hideSidebar:!sidebar.opened}">
-        <div class="sidebar-wrapper">
-            <sidebar class="sidebar-container"></sidebar>
-        </div>
-        <div class="main-container">
-            <navbar></navbar>
-            <app-main></app-main>
-        </div>
+  <div class="app-wrapper" :class="{hideSidebar:!sidebar.opened}">
+    <div class="sidebar-wrapper">
+      <sidebar class="sidebar-container"></sidebar>
     </div>
+    <div class="main-container">
+      <navbar></navbar>
+      <app-main></app-main>
+    </div>
+  </div>
 </template>
 
-<script>
-    import { Navbar, Sidebar, AppMain } from '@/views/layout';
 
-    export default {
-      name: 'layout',
-      components: {
-        Navbar,
-        Sidebar,
-        AppMain
-      },
-      computed: {
-        sidebar() {
-          return this.$store.state.app.sidebar;
-        }
-      }
+<script>
+import { Navbar, Sidebar, AppMain } from '@/views/layout'
+
+export default {
+  name: 'layout',
+  components: {
+    Navbar,
+    Sidebar,
+    AppMain
+  },
+  computed: {
+    sidebar() {
+      return this.$store.state.app.sidebar
     }
+  }
+}
 </script>
 
 <style rel="stylesheet/scss" lang="scss" scoped>
