@@ -1,25 +1,24 @@
 <template>
-    <el-menu mode="vertical" theme="dark" :default-active="$route.path">
-        <sidebar-item :routes='permission_routers'></sidebar-item>
-    </el-menu>
+  <el-menu mode="vertical" theme="dark" :default-active="$route.path">
+    <sidebar-item :routes='permission_routers'></sidebar-item>
+  </el-menu>
 </template>
 
 <script>
-    import { mapGetters } from 'vuex';
-    import SidebarItem from './SidebarItem';
-    export default {
-      components: { SidebarItem },
-      computed: {
-        ...mapGetters([
-          'permission_routers'
-        ])
-      }
-    }
+import { mapGetters } from 'vuex'
+import SidebarItem from './SidebarItem'
+export default {
+  components: { SidebarItem },
+  computed: {
+    ...mapGetters([
+      'permission_routers'
+    ])
+  }
+}
 </script>
 
 <style rel="stylesheet/scss" lang="scss" scoped>
 .el-menu {
-    min-height: 100%;
+  min-height: 100%;
 }
-
 </style>
