@@ -1,7 +1,7 @@
-export { default as Navbar } from './Navbar';
+export { default as Navbar } from './Navbar'
 
-export { default as Sidebar } from './Sidebar';
+export { default as Sidebar } from './Sidebar'
 
-export { default as Levelbar } from './Levelbar';
+export { default as Levelbar } from './Levelbar'
 
-export { default as AppMain } from './AppMain';
+export { default as AppMain } from './AppMain'
